requirements.txt
# Django Core
Django==6.0.4
djangorestframework==3.17.1
djangorestframework-simplejwt==5.4.0
django-cors-headers==4.3.1

# Database
psycopg[binary]==3.3.3
dj-database-url==2.1.0

# Async & WebSocket
channels==4.0.0
channels-redis==4.2.0
daphne==4.1.2

# Celery
celery==5.3.6
redis==5.0.3

# Image & QR
Pillow==11.0.0
qrcode==7.4.2

# Storage
django-storages==1.14.2
boto3==1.34.84

# Filtering & Search
django-filter==24.2

# API Documentation
drf-spectacular==0.27.2

# Security & Auth
bcrypt==4.1.2

# Utils
python-dotenv==1.0.1
requests==2.31.0

# Production Server
gunicorn==21.2.0
whitenoise==6.6.0
